+++ src/firecracker/examples/uffd/fault_all_handler.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2024 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 //! Provides functionality for a userspace page fault handler
@@ -10,7 +10,7 @@ mod uffd_utils;
 use std::fs::File;
 use std::os::unix::net::UnixListener;
 
-use uffd_utils::{MemPageState, Runtime, UffdHandler};
+use uffd_utils::{Runtime, UffdHandler};
 use utils::get_page_size;
 
 fn main() {
@@ -27,25 +27,23 @@ fn main() {
     // Populate a single page from backing memory file.
     // This is just an example, probably, with the worst-case latency scenario,
     // of how memory can be loaded in guest RAM.
-    let len = get_page_size().unwrap();
+    let len = get_page_size().unwrap(); // page size does not matter, we fault in everything on the first fault
 
     let mut runtime = Runtime::new(stream, file);
-    runtime.run(|uffd_handler: &mut UffdHandler| {
+    runtime.run(len, |uffd_handler: &mut UffdHandler| {
         // Read an event from the userfaultfd.
         let event = uffd_handler
             .read_event()
             .expect("Failed to read uffd_msg")
             .expect("uffd_msg not ready");
 
-        // We expect to receive either a Page Fault or Removed
-        // event (if the balloon device is enabled).
         match event {
-            userfaultfd::Event::Pagefault { addr, .. } => uffd_handler.serve_pf(addr.cast(), len),
-            userfaultfd::Event::Remove { start, end } => uffd_handler.update_mem_state_mappings(
-                start as u64,
-                end as u64,
-                &MemPageState::Removed,
-            ),
+            userfaultfd::Event::Pagefault { .. } => {
+                for region in uffd_handler.mem_regions.clone() {
+                    uffd_handler
+                        .serve_pf(region.mapping.base_host_virt_addr as _, region.mapping.size)
+                }
+            }
             _ => panic!("Unexpected event on userfaultfd"),
         }
     });