+++ src/vmm/src/rpc_interface.rs
@@ -1069,6 +1069,7 @@ mod tests {
                 smt: value.vm_config.smt,
                 cpu_template: StaticCpuTemplate::from(&value.vm_config.cpu_template),
                 boot_source: value.boot_source_config().clone(),
+                huge_pages: value.vm_config.huge_pages,
             }
         }
     }