+++ src/vmm/src/vmm_config/machine_config.rs
@@ -3,6 +3,8 @@
 use std::fmt::Debug;
 
 use serde::{Deserialize, Serialize};
+use utils::kernel_version;
+use utils::kernel_version::KernelVersion;
 
 use crate::cpu_config::templates::{CpuTemplateType, CustomCpuTemplate, StaticCpuTemplate};
 
@@ -18,7 +20,7 @@ pub const MAX_SUPPORTED_VCPUS: u8 = 32;
 pub enum VmConfigError {
     /// The memory size (MiB) is smaller than the previously set balloon device target size.
     IncompatibleBalloonSize,
-    /// The memory size (MiB) is invalid.
+    /// The memory size (MiB) is either 0, or not a multiple of the configured page size.
     InvalidMemorySize,
     /// The number of vCPUs must be greater than 0, less than {MAX_SUPPORTED_VCPUS:} and must be 1 or an even number if SMT is enabled.
     InvalidVcpuCount,
@@ -27,6 +29,70 @@ pub enum VmConfigError {
     /// Enabling simultaneous multithreading is not supported on aarch64.
     #[cfg(target_arch = "aarch64")]
     SmtNotSupported,
+    /// Could not determine host kernel version when checking hugetlbfs compatibility
+    KernelVersion,
+    /// Firecracker's hugetlbfs support requires at least host kernel 5.10.
+    HugetlbfsNotSupported,
+    /// Firecracker's huge pages support is incompatible with memory ballooning.
+    BalloonAndHugePages,
+    /// Firecracker's huge pages support is incompatible with initrds.
+    InitrdAndHugePages,
+}
+
+// We cannot do a `KernelVersion(kernel_version::Error)` variant because `kernel_version::Error`
+// does not implement `PartialEq, Eq` (due to containing an io error).
+impl From<kernel_version::Error> for VmConfigError {
+    fn from(_: kernel_version::Error) -> Self {
+        VmConfigError::KernelVersion
+    }
+}
+
+/// Describes the possible (huge)page configurations for a microVM's memory.
+#[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
+pub enum HugePageConfig {
+    /// Do not use hugepages, e.g. back guest memory by 4K
+    #[default]
+    None,
+    /// Back guest memory by 2MB hugetlbfs pages
+    #[serde(rename = "2M")]
+    Hugetlbfs2M,
+}
+
+impl HugePageConfig {
+    /// Checks whether the given memory size (in MiB) is valid for this [`HugePageConfig`], e.g.
+    /// whether it is a multiple of the page size
+    fn is_valid_mem_size(&self, mem_size_mib: usize) -> bool {
+        let divisor = match self {
+            // Any integer memory size expressed in MiB will be a multiple of 4096KiB.
+            HugePageConfig::None => 1,
+            HugePageConfig::Hugetlbfs2M => 2,
+        };
+
+        mem_size_mib % divisor == 0
+    }
+
+    /// Returns the flags required to pass to `mmap`, in addition to `MAP_ANONYMOUS`, to
+    /// create a mapping backed by huge pages as described by this [`HugePageConfig`].
+    pub fn mmap_flags(&self) -> libc::c_int {
+        match self {
+            HugePageConfig::None => 0,
+            HugePageConfig::Hugetlbfs2M => libc::MAP_HUGETLB | libc::MAP_HUGE_2MB,
+        }
+    }
+
+    /// Returns `true` iff this [`HugePageConfig`] describes a hugetlbfs-based configuration.
+    pub fn is_hugetlbfs(&self) -> bool {
+        matches!(self, HugePageConfig::Hugetlbfs2M)
+    }
+}
+
+impl From<HugePageConfig> for Option<memfd::HugetlbSize> {
+    fn from(value: HugePageConfig) -> Self {
+        match value {
+            HugePageConfig::None => None,
+            HugePageConfig::Hugetlbfs2M => Some(memfd::HugetlbSize::Huge2MB),
+        }
+    }
 }
 
 /// Struct used in PUT `/machine-config` API call.
@@ -46,6 +112,9 @@ pub struct MachineConfig {
     /// Enables or disables dirty page tracking. Enabling allows incremental snapshots.
     #[serde(default)]
     pub track_dirty_pages: bool,
+    /// Configures what page size Firecracker should use to back guest memory.
+    #[serde(default)]
+    pub huge_pages: HugePageConfig,
 }
 
 impl Default for MachineConfig {
@@ -78,6 +147,9 @@ pub struct MachineConfigUpdate {
     /// Enables or disables dirty page tracking. Enabling allows incremental snapshots.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub track_dirty_pages: Option<bool>,
+    /// Configures what page size Firecracker should use to back guest memory.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub huge_pages: Option<HugePageConfig>,
 }
 
 impl MachineConfigUpdate {
@@ -97,6 +169,7 @@ impl From<MachineConfig> for MachineConfigUpdate {
             smt: Some(cfg.smt),
             cpu_template: cfg.cpu_template,
             track_dirty_pages: Some(cfg.track_dirty_pages),
+            huge_pages: Some(cfg.huge_pages),
         }
     }
 }
@@ -114,6 +187,8 @@ pub struct VmConfig {
     pub cpu_template: Option<CpuTemplateType>,
     /// Enables or disables dirty page tracking. Enabling allows incremental snapshots.
     pub track_dirty_pages: bool,
+    /// Configures what page size Firecracker should use to back guest memory.
+    pub huge_pages: HugePageConfig,
 }
 
 impl VmConfig {
@@ -148,8 +223,9 @@ impl VmConfig {
         }
 
         let mem_size_mib = update.mem_size_mib.unwrap_or(self.mem_size_mib);
+        let page_config = update.huge_pages.unwrap_or(self.huge_pages);
 
-        if mem_size_mib == 0 {
+        if mem_size_mib == 0 || !page_config.is_valid_mem_size(mem_size_mib) {
             return Err(VmConfigError::InvalidMemorySize);
         }
 
@@ -159,12 +235,17 @@ impl VmConfig {
             Some(other) => Some(CpuTemplateType::Static(other)),
         };
 
+        if page_config.is_hugetlbfs() && KernelVersion::get()? < KernelVersion::new(4, 16, 0) {
+            return Err(VmConfigError::HugetlbfsNotSupported);
+        }
+
         Ok(VmConfig {
             vcpu_count,
             mem_size_mib,
             smt,
             cpu_template,
             track_dirty_pages: update.track_dirty_pages.unwrap_or(self.track_dirty_pages),
+            huge_pages: page_config,
         })
     }
 }
@@ -177,6 +258,7 @@ impl Default for VmConfig {
             smt: false,
             cpu_template: None,
             track_dirty_pages: false,
+            huge_pages: HugePageConfig::None,
         }
     }
 }
@@ -189,6 +271,31 @@ impl From<&VmConfig> for MachineConfig {
             smt: value.smt,
             cpu_template: value.cpu_template.as_ref().map(|template| template.into()),
             track_dirty_pages: value.track_dirty_pages,
+            huge_pages: value.huge_pages,
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use utils::kernel_version::KernelVersion;
+
+    use crate::vmm_config::machine_config::{
+        HugePageConfig, MachineConfigUpdate, VmConfig, VmConfigError,
+    };
+
+    #[test]
+    fn test_hugetlbfs_not_supported_4_14() {
+        if KernelVersion::get().unwrap() < KernelVersion::new(4, 16, 0) {
+            let base_config = VmConfig::default();
+            let update = MachineConfigUpdate {
+                huge_pages: Some(HugePageConfig::Hugetlbfs2M),
+                mem_size_mib: Some(1024),
+                ..Default::default()
+            };
+
+            let err = base_config.update(&update).unwrap_err();
+            assert_eq!(err, VmConfigError::HugetlbfsNotSupported)
         }
     }
 }