+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to
   `VcpuExit::MmioRead`, `VcpuExit::MmioWrite`, `VcpuExit::IoIn` and
   `VcpuExit::IoOut`. The average for these VM exits is not emitted since it can
   be deduced from the available emitted metrics.
+- [#4360](https://github.com/firecracker-microvm/firecracker/pull/4360): Added
+  dev-preview support for backing a VM's guest memory by 2M hugetlbfs pages.
+  Please see the [documentation](docs/hugepages.md) for more information.
 
 ### Changed